dmn-tck-it/dmn-tck-it-python-translator/requirements.txt
isodate==0.6.1
jdmn-python-runtime==0.0.4
